(PUP-10432) Fix yum provider check_updates.
RedHat RHBA-2020:1028 changed the output of `yum check-update`
when subscription manager is enabled(adding extra blank lines) breaking
the `check_updates` method of `yum provider`.

This commit changes how we parse the output of `yum check-update`
in order not to depend on the first blank line occured but rather
split by blank lines, and check on each chunk to see
if we can extract package information.